requirements.txt
python-escpos[all]
python-dotenv
pyyaml
coloredlogs
asyncio
websockets